fix: create proper redirect for non-locale paths
When the path didn’t have a locale but the `localePrefix` was set to
`always` we let next-intl redirect the user to
`/${defaultLocale}/${path}`. On visit we then rewrote that to
`/${defaultPathLocale}/${path}` and redirect the client again.

This makes sure we do the redirect ourselves skipping an extra redirect